new density and corrected cell counting
Introduced new density measure: area covered by signal. Corrected the
way the signal is identified, which is now based on corrected image.

Changed the counting of the cells for the overlay image by saving,
which cells are used for analysis.

Added Otsu 2D function, just in case I might use it in future for
thresholding.

Removed separate density outputs, as they are repeated in summary.